pyproject.toml
[project]
name = "alphastrike"
version = "2.0.0"
description = "Autonomous ML-based crypto trading system for WEEX perpetual futures"
readme = "README.md"
requires-python = ">=3.12"
license = {text = "MIT"}
authors = [
    {name = "Bowen Li"}
]

dependencies = [
    # Async
    "aiohttp>=3.9.0",
    "aiosqlite>=0.19.0",
    "websockets>=12.0",

    # ML - Gradient Boosting
    "xgboost>=2.0.0",
    "lightgbm>=4.0.0",
    "scikit-learn>=1.4.0",
    "scipy>=1.11.0",

    # ML - Deep Learning
    "torch>=2.1.0",
    "numpy>=1.26.0",
    "pandas>=2.1.0",

    # Configuration
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "python-dotenv>=1.0.0",

    # Utilities
    "structlog>=23.2.0",
    "orjson>=3.9.0",
    "joblib>=1.3.0",

    # Technical Analysis
    "ta>=0.11.0",

    # CLI
    "click>=8.1.0",
    "rich>=13.0.0",  # Pretty terminal output
]

[project.scripts]
alphastrike = "src.cli.main:main"

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.1.0",
    "pyright>=1.1.340",
    "ruff>=0.1.6",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.pyright]
pythonVersion = "3.12"
typeCheckingMode = "basic"
reportMissingImports = true
reportMissingTypeStubs = false
include = ["src"]
exclude = ["tests", "**/__pycache__"]
venvPath = "."
venv = ".venv"

[tool.ruff]
target-version = "py312"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]
ignore = [
    "E501",   # Line too long (handled by formatter)
    "N802",   # Function name should be lowercase (config constants: MAX_LEVERAGE)
    "N803",   # Argument name should be lowercase (ML convention: X, X_train)
    "N806",   # Variable in function should be lowercase (ML convention: X, X_train)
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]
